+++ Source-Codes/IERSystem/IERSystem/Areas/QuanLySoNhanMau/Controllers/CacSoNhanMauController.cs
@@ -28,7 +28,7 @@ namespace IERSystem.Areas.QuanLySoNhanMau.Controllers
             if (ModelState.IsValid)
             {
                 var today = DateTime.Now;
-                cacsonhanmau_inp.CreatedBy = db.Users.Find((int)Session["loggedID"]);
+                cacsonhanmau_inp.CreatedBy = db.Users.Find(Session["loggedID"]);
                 cacsonhanmau_inp.UserId = cacsonhanmau_inp.CreatedBy.Id;
                 cacsonhanmau_inp.CreatedDate = DateTime.Now;
                 cacsonhanmau_inp.Nam = today.Year;
